Fixed minor issue with affinity determination with few subtasks Fixed an issue with alternating use of address space as RO and WO

+++ source/code/src/pmPreprocessorTask.cpp
@@ -129,7 +129,7 @@ pmStatus preprocessorTask_cpuCallback(pmTaskInfo pTaskInfo, pmDeviceInfo pDevice
                 lSelectiveSubtasks = true;
                 
                 ulong lSubtasksPerMachine = pTaskInfo.subtaskCount / lTaskConf->machinesCount;
-                float lSubtaskSpan = lTaskConf->originalTaskSubtasks / lSubtasksPerMachine;
+                float lSubtaskSpan = (float)lTaskConf->originalTaskSubtasks / lSubtasksPerMachine;
                 
                 EXCEPTION_ASSERT(lSubtaskSpan > 1.0);
